added mixins for the simplicity

+++ store/views.py
@@ -1,43 +1,30 @@
 from django.shortcuts import get_object_or_404
-from rest_framework.decorators import api_view
+from django.db.models import Count
 from rest_framework.response import Response
-from rest_framework.request import Request
 from rest_framework import status
 from .models import Collection, Product
 from .serializers import ProductSerializer, CollectionSerializer
-from rest_framework.views import APIView
-from rest_framework.mixins import ListModelMixin, CreateModelMixin
+from rest_framework.generics import ListCreateAPIView, RetrieveUpdateDestroyAPIView
+
 
 # Create your views here.
 
-class ProductList(APIView):
-        def get(self, request):
-            query_set = Product.objects.select_related('collection').all()
-            serializer = ProductSerializer(query_set, many=True)
-            return Response(serializer.data)
-
-        def post(self, request):
-                serializer = ProductSerializer(data=request.data)
-                serializer.is_valid(raise_exception=True)
+class ProductList(ListCreateAPIView):
+        queryset = Product.objects.select_related('collection').all()
+        serializer_class = ProductSerializer
         
-                serializer.save()
-                return Response(serializer.data, status=status.HTTP_201_CREATED)
+        def get_serializer_context(self):
+                return {'request': self.request }
 
 
-class ProductDetails(APIView):
-        def get(self, request, id):
-                product = get_object_or_404(Product, pk=id)
-                serializer = ProductSerializer(product)
-                return Response(serializer.data)
+class ProductDetails(RetrieveUpdateDestroyAPIView):
+        queryset = Product.objects.all()
+        serializer_class = ProductSerializer
 
-        def put(self, request, id):
-                product = get_object_or_404(Product, pk=id)
-                serializer = ProductSerializer(product,data=request.data)
-                serializer.is_valid(raise_exception=True)
-                
-                serializer.save()
-                return Response(serializer.data, status=status.HTTP_201_CREATED)
+        def get_serializer_context(self):
+                return {'request': self.request }
         
+        # HERE I CUSTOMIZED THE DELETE METHOD SO THAT I CAN ADD MY EXTRA LOGIC
         def delete(self, request, id):
                 product = get_object_or_404(Product, pk=id)
                 if product.orderitem_set.count() > 0:
@@ -47,33 +34,21 @@ class ProductDetails(APIView):
                 return Response(status=status.HTTP_204_NO_CONTENT)
 
 
-class CollectionList(APIView):
-        def get(self, request):
-                collections = Collection.objects.all()
-                serializer = CollectionSerializer(collections, many=True)
-                return Response(serializer.data)
-        
-        def post(self, request):
-                serializer = CollectionSerializer(data=request.data)
-                serializer.is_valid(raise_exception=True)
-                serializer.save()
-                return Response(serializer.data, status=status.HTTP_201_CREATED)
+class CollectionList(ListCreateAPIView):
+        queryset = Collection.objects.annotate(products_count=Count('products')).all()
+        serializer_class = CollectionSerializer
+
+        def get_serializer_context(self):
+                return {'request': self.request }
                 
+                
+class CollectionDetails(RetrieveUpdateDestroyAPIView):
+        queryset = Collection.objects.annotate(products_count=Count('products'))
+        serializer_class = CollectionSerializer
 
-class CollectionDetails(APIView):
-        def get(self, request, id):
-                collection = get_object_or_404(Collection, pk=id)
-                serializer = CollectionSerializer(collection)
-                return Response(serializer.data)
-        
-        def put(self, request, id):
-                collection = get_object_or_404(Collection, pk=id)
-                serializer = CollectionSerializer(collection, data=request.data)
-                serializer.is_valid(raise_exception=True)
-                serializer.save()
-                return Response(serializer.data, status=status.HTTP_201_CREATED)
-
-        def delete(self, request, id):
-                collection = get_object_or_404(Collection, pk=id)
+        def delete(self, request, pk):
+                collection: Collection = get_object_or_404(Collection, pk=pk)
+                if collection.products.count() > 0:
+                        return Response({'error': 'Collection cannot be deleted, it is associated with products'}, status=status.HTTP_406_NOT_ACCEPTABLE)
                 collection.delete()
                 return Response(status=status.HTTP_204_NO_CONTENT)
